Fixed zondax-ledger-icp dependencies
