+++ package.json
@@ -17,7 +17,7 @@
 	"devDependencies": {
 		"@dfinity/identity-secp256k1": "^2.2.0",
 		"@dfinity/internet-identity-playwright": "^0.0.2",
-		"@playwright/test": "^1.46.0",
+		"@playwright/test": "^1.52.0",
 		"@sveltejs/adapter-static": "^2.0.3",
 		"@sveltejs/kit": "^1.30.4",
 		"@sveltejs/vite-plugin-svelte": "^2.4.2",
@@ -26,7 +26,7 @@
 		"dotenv": "^16.4.5",
 		"events": "^3.3.0",
 		"fast-check": "^3.21.0",
-		"playwright": "^1.46.0",
+		"playwright": "^1.52.0",
 		"prettier": "^3.1.1",
 		"prettier-plugin-svelte": "^3.1.2",
 		"svelte": "^4.2.19",
@@ -55,7 +55,8 @@
 		"@slide-computer/signer-agent": "^3.16.0",
 		"@slide-computer/signer-extension": "^3.16.0",
 		"@slide-computer/signer-web": "^3.16.0",
-		"@zondax/ledger-icp": "^0.6.1",
+		"@zondax/ledger-icp": "^3.2.8",
+		"@zondax/ledger-js": "^1.2.0",
 		"flowbite-svelte": "^0.48.4",
 		"process": "^0.11.10",
 		"qr-creator": "^1.0.0"
